Fix merge conflict

+++ src/pages/Categories/CategoriesList/CategoriseListPage.jsx
@@ -1,9 +1,9 @@
+import { useContext } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import ListPage from "../../../components/shared/ListPage/ListPage";
 import tblFields from "./tblFields.json";
-import { useContext } from "react";
 import AuthContext from "../../../store/auth-contex";
-// import ModalContent from "../ModalContent";
+import ModalContent from "./ModalContent";
 import { toast } from "react-toastify";
 
 const CategoriesListPage = () => {
@@ -24,35 +24,43 @@ const CategoriesListPage = () => {
 
     const customActions = {
         delete: {
-            clickHandler: {
-                type: "dialog_delete",
-                fnc: (rowData, handleDeleteModalData) => {
-                    return {
-                        show: true,
-                        id: rowData.id,
-                        mutate: null,
-                        // children: (
-                        //   <ModalContent apiPath={`admin/category-product/categories/${rowData.id}`} rowData={rowData} handleDeleteModalData={handleDeleteModalData} />
-                        // )
-                    };
-                },
+          clickHandler: {
+            type: 'dialog_delete',
+            fnc: (rowData, handleDeleteModalData) => {
+              return {
+                show: true,
+                id: rowData.id,
+                mutate: null,
+                children: (
+                  <ModalContent apiPath={`admin/category-product/categories/message/${rowData.id}`} rowData={rowData} handleDeleteModalData={handleDeleteModalData} />
+                )
+              };
             },
-            deleteClickHandler: {
-                type: "dialog_delete",
-                fnc: (rowData, data) => {
-                    api.delete(`admin/category-product/categories/${rowData.id}`)
-                        .then(() => toast.success("Zapis je uspešno obrisan"))
-                        .catch(() => toast.warning("Došlo je do greške prilikom brisanja"));
-
-                    return {
-                        show: false,
-                        id: rowData.id,
-                        mutate: 1,
-                    };
-                },
-            },
+          },
+          deleteClickHandler: {
+            type: 'dialog_delete',
+            fnc: (rowData, deleteModalData) => {
+    
+              if(deleteModalData.all_fill) {
+                api.delete(`admin/category-product/categories/confirm/${rowData.id}`, deleteModalData)
+                  .then(() => toast.success("Zapis je uspešno obrisan"))
+                  .catch(() => toast.warning("Došlo je do greške prilikom brisanja"));
+    
+                return {
+                  show: false,
+                  id: rowData.id,
+                  mutate: 1,
+                };
+              } else {
+                toast.warning("Potrebno je da povežete sve opcije koje se brišu");
+                return false;
+              }
+              
+            }
+          },
         },
-    };
+      };
+    
 
     return (
         <ListPage
